Push VerifyObject to ClrHeap
